refactor: rename files to PascalCase/kebab-case, add barrel exports, move Card test
