+++ src/components/features/about/CertificationItem.astro
@@ -1,6 +1,5 @@
 ---
-import Card from "@/components/ui/card/Card.astro";
-import CardContent from "@/components/ui/card/CardContent.astro";
+import { Card, CardContent } from "@/components/ui/card";
 import type { Certification } from "@/types";
 
 interface Props {
