+++ src/components/features/blog/PostPreview.astro
@@ -2,7 +2,7 @@
 import type { CollectionEntry } from "astro:content";
 import PostMetadata from "@/components/common/PostMetadata.astro";
 import Badge from "@/components/ui/Badge.astro";
-import Card from "@/components/ui/card/Card.astro";
+import { Card } from "@/components/ui/card";
 import { cn, getPostUrl } from "@/lib/utils";
 
 type Props = {
