+++ src/components/features/blog/RelatedPost.astro
@@ -1,9 +1,6 @@
 ---
 import PostMetadata from "@/components/common/PostMetadata.astro";
-import Card from "@/components/ui/card/Card.astro";
-import CardFooter from "@/components/ui/card/CardFooter.astro";
-import CardHeader from "@/components/ui/card/CardHeader.astro";
-import CardTitle from "@/components/ui/card/CardTitle.astro";
+import { Card, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
 import Heading from "@/components/ui/Heading.astro";
 import { getPostUrl } from "@/lib/utils";
 import type { CollectionPosts } from "@/types";
