+++ src/components/features/home/FeaturedProjectItem.astro
@@ -1,6 +1,6 @@
 ---
 import Badge from "@/components/ui/Badge.astro";
-import Card from "@/components/ui/card/Card.astro";
+import { Card } from "@/components/ui/card";
 import { cn } from "@/lib/utils";
 import type { Project } from "@/types";
