+++ src/components/shell/Header.astro
@@ -3,7 +3,7 @@ import { Image } from "astro:assets";
 import logoDark from "@/assets/images/logo-dark.svg";
 import logoLight from "@/assets/images/logo-light.svg";
 import NavLink from "@/components/shell/NavLink.astro";
-import Search from "@/components/ui/search/Search.svelte";
+import { Search } from "@/components/ui/search";
 import ThemeToggle from "@/components/ui/ThemeToggle.astro";
 import siteConfig from "@/config/site-config";
