+++ src/components/ui/InfoPanel.astro
@@ -1,8 +1,5 @@
 ---
-import Card from "@/components/ui/card/Card.astro";
-import CardContent from "@/components/ui/card/CardContent.astro";
-import CardHeader from "@/components/ui/card/CardHeader.astro";
-import CardTitle from "@/components/ui/card/CardTitle.astro";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { cn } from "@/lib/utils";
 import Stack from "./primitives/Stack.astro";
