+++ src/components/ui/ThemeToggle.astro
@@ -16,7 +16,7 @@ import Button from "@/components/ui/Button.astro";
 </theme-toggle>
 
 <script>
-	import { rootInDarkMode } from "@/lib/utils/domElement";
+	import { rootInDarkMode } from "@/lib/utils/dom-element";
 
 	class ThemeToggle extends HTMLElement {
 		constructor() {
