+++ src/components/ui/card/index.ts
@@ -1,0 +1,6 @@
+export { default as Card } from "./Card.astro";
+export { default as CardContent } from "./CardContent.astro";
+export { default as CardDescription } from "./CardDescription.astro";
+export { default as CardFooter } from "./CardFooter.astro";
+export { default as CardHeader } from "./CardHeader.astro";
+export { default as CardTitle } from "./CardTitle.astro";
