+++ src/components/ui/search/index.ts
@@ -1,0 +1,6 @@
+export { default as Kbd } from "./Kbd.svelte";
+export { default as Search } from "./Search.svelte";
+export { default as SearchBar } from "./SearchBar.svelte";
+export { default as SearchModal } from "./SearchModal.svelte";
+export { default as SearchResultItem } from "./SearchResultItem.svelte";
+export { default as SearchTrigger } from "./SearchTrigger.svelte";
