+++ src/pages/blog/series/index.astro
@@ -1,7 +1,7 @@
 ---
 import type { CollectionEntry } from "astro:content";
 import BlogTabs from "@/components/features/blog/BlogTabs.astro";
-import Card from "@/components/ui/card/Card.astro";
+import { Card } from "@/components/ui/card";
 import Heading from "@/components/ui/Heading.astro";
 import BaseLayout from "@/layouts/BaseLayout.astro";
 import { getAllSeries } from "@/lib/contents";
